requirements.txt
dnspython>=2.8.0
requests>=2.32.0
